ac3: clippy sweep — unnecessary_cast, identity_op, needless_range_loop
- `src/bsi.rs`: drop `as u32` cast on `BitReader::read_u32`
  (clippy::unnecessary_cast).
- `src/syncinfo.rs`: drop `(0 << 6) | N` / `0 | N` constructions in the
  tests — the shifted field is always zero so the identity operation
  carries no information (clippy::identity_op).
- `tests/downmix_51.rs`: replace an index loop filling a whole row with
  `slice::fill` (clippy::needless_range_loop).

Pure style, no behaviour change. Transient-fixture PSNR stays at
15.55 dB.

+++ src/bsi.rs
@@ -142,7 +142,7 @@ pub fn parse(data: &[u8]) -> Result<Bsi> {
     // addbsi — up to 64 bytes of trailing info we can safely skip.
     let addbsie = br.read_u32(1)? != 0;
     if addbsie {
-        let addbsil = br.read_u32(6)? as u32; // 0..=63, meaning 1..=64 bytes
+        let addbsil = br.read_u32(6)?; // 0..=63, meaning 1..=64 bytes
         let nbits = (addbsil + 1) * 8;
         br.skip(nbits)?;
     }
